fix(ci): pin ruff and clear its 0.16 default-ruleset findings
CI installs `ruff` unpinned via the `dev` extra. Ruff 0.16.0 widened its
*default* rule set well beyond the old E4/E7/E9/F — isort, bugbear,
pyupgrade, flake8-comprehensions, RUF, TRY, FURB are now on by default —
so `ruff check src tests` went from clean to 64 errors with no change on
our side. Nothing in b3dd0ee caused it.

Two halves:

- pyproject: pin `ruff>=0.16,<0.17` so the lint gate only moves when we
  move it, and state the rule set we actually want. C408 is ignored on
  purpose (dict(flux=..., err=...) reads better than a quoted literal for
  the keyword-style records the backends pass around), and isort is told
  that `fixtures` is the tests' own helper package, not third-party.

- source/tests: sorted import blocks and `__all__`, dropped the unused
  `select_psf_native` import and a stale `noqa: E731`, `lru_cache(None)`
  -> `cache`, unquoted two self-referential annotations, generator ->
  set comprehension, `sorted(...)[0]` -> `min(...)`, dropped `int()`
  around `round()`/`math.floor()` (both already return int), and dropped
  the redundant exception object from two `logger.exception` calls.

The four deliberate blind `except Exception` guards (unusable WCS, failed
Background2D, Data Lab error detail, unreadable cutout header) keep the
behaviour and carry an explicit `noqa: BLE001` with the reason, so BLE001
still catches accidental ones.

Verified: `ruff check src tests` clean under 0.16.0, `pytest -q` 65 passed.

+++ src/spherex_photometry/backends/cpu_psf.py
@@ -72,7 +72,7 @@ class OversampledPixelizedPSF(PixelizedPSF):
             # shift by the sub-pixel offset at oversampled resolution, then sum
             # each k x k block. This keeps the PSF pixel-integrated at native
             # scale rather than point-sampled.
-            k = int(round(factor))
+            k = round(factor)
             target_h = self.nativeH * k
             target_w = self.nativeW * k
             h, w = img.shape
@@ -85,8 +85,8 @@ class OversampledPixelizedPSF(PixelizedPSF):
             target_center_y = crop_y0 + (target_h - 1) / 2.0
             desired_x = target_center_x + dx * k
             desired_y = target_center_y + dy * k
-            pw = int(round(desired_x - (w // 2)))
-            ph = int(round(desired_y - (h // 2)))
+            pw = round(desired_x - (w // 2))
+            ph = round(desired_y - (h // 2))
             pw = max(0, min(canvas_w - w, pw))
             ph = max(0, min(canvas_h - h, ph))
             pad_img = np.zeros((canvas_h, canvas_w), dtype=img.dtype)
